# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: bloq5 — Plateforme de gestion immobilière
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List properties with filters
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [house, apartment, co-living, commercial, office, industrial]
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: bedrooms
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyListResponse"

    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a new property listing (owner/manager)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyBody"
      responses:
        "201":
          description: Created property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
...
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get property details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProperty
      tags: [properties]
      summary: Update property (owner/manager)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePropertyBody"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"

    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete property (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /properties/{id}/available-dates:
    get:
      operationId: getPropertyAvailableDates
      tags: [properties]
      summary: Get available visit dates for a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Available dates
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: date

  # --- RENTAL REQUESTS ---
...
  /favorites:
    get:
      operationId: listFavorites
      tags: [favorites]
      summary: Get user favorites
      responses:
        "200":
          description: Favorite properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"

    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add property to favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFavoriteBody"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /favorites/{propertyId}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove property from favorites
      parameters:
        - in: path
          name: propertyId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # --- VISITS ---
...
  /messages/conversation/{requestId}:
    get:
      operationId: getConversationMessages
      tags: [messages]
      summary: Get messages for a rental request conversation
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a rental request conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /conversations/{requestId}:
    delete:
      operationId: endConversation
      tags: [messages]
      summary: End a conversation (owner/manager)
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Conversation ended

  # --- DASHBOARD ---
...
  /managers:
    get:
      operationId: listManagers
      tags: [managers]
      summary: List managers assigned by owner
      responses:
        "200":
          description: Managers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Manager"

    post:
      operationId: addManager
      tags: [managers]
      summary: Add a manager (requires owner auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddManagerBody"
      responses:
        "201":
          description: Manager added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manager"
  /managers/{id}:
    delete:
      operationId: removeManager
      tags: [managers]
      summary: Remove a manager
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # --- SUBSCRIPTIONS ---
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    Property:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [house, apartment, co-living, commercial, office, industrial]
        address:
          type: string
        city:
          type: string
        country:
          type: string
        price:
          type: number
        bedrooms:
          type: ["integer", "null"]
        bathrooms:
          type: ["integer", "null"]
        area:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        virtualTourUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, rented, unavailable]
        isFeatured:
          type: boolean
        views:
          type: integer
        requestsCount:
          type: integer
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - type
        - address
        - city
        - country
        - price
        - images
        - status
        - isFeatured
        - views
        - requestsCount
        - ownerId
        - createdAt
    PropertyDetail:
      allOf:
        - $ref: "#/components/schemas/Property"
        - type: object
          properties:
            amenities:
              type: array
              items:
                type: string
            floorPlan:
              type: ["string", "null"]
            nearbyPlaces:
              type: array
              items:
                type: string
            ownerName:
              type: string
            ownerEmail:
              type: string
            ownerPhone:
              type: ["string", "null"]
            availableDates:
              type: array
              items:
                type: string
          required:
            - amenities
            - nearbyPlaces
            - ownerName
            - ownerEmail
            - availableDates
    PropertyListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    CreatePropertyBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [house, apartment, co-living, commercial, office, industrial]
        address:
          type: string
        city:
          type: string
        country:
          type: string
        price:
          type: number
        bedrooms:
          type: ["integer", "null"]
        bathrooms:
          type: ["integer", "null"]
        area:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        virtualTourUrl:
          type: ["string", "null"]
        amenities:
          type: array
          items:
            type: string
        availableDates:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
      required:
        - title
        - description
        - type
        - address
        - city
        - country
        - price
        - images
    UpdatePropertyBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        status:
          type: string
          enum: [available, rented, unavailable]
        images:
          type: array
          items:
            type: string
        virtualTourUrl:
          type: ["string", "null"]
        amenities:
          type: array
          items:
            type: string
        availableDates:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
...
    Message:
      type: object
      properties:
        id:
          type: integer
        requestId:
          type: integer
        senderId:
          type: string
        senderName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
      required:
        - id
        - requestId
        - senderId
        - senderName
        - content
        - createdAt
        - expiresAt
    SendMessageBody:
      type: object
      properties:
        requestId:
          type: integer
        content:
          type: string
      required: [requestId, content]
...
    Manager:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: string
        managerEmail:
          type: string
        managerName:
          type: string
        managerPhone:
          type: ["string", "null"]
        permissions:
          type: array
          items:
            type: string
        assignedProperties:
          type: array
          items:
            type: integer
        status:
          type: string
          enum: [pending, verified, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ownerId
        - managerEmail
        - managerName
        - permissions
        - assignedProperties
        - status
        - createdAt
    AddManagerBody:
      type: object
      properties:
        managerEmail:
          type: string
        managerName:
          type: string
        managerPhone:
          type: ["string", "null"]
        ownerEmail:
          type: string
        ownerPhone:
          type: ["string", "null"]
        assignedProperties:
          type: array
          items:
            type: integer
      required: [managerEmail, managerName, ownerEmail]
...