# lib/api-zod/src/generated/types/createPropertyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * bloq5 — Plateforme de gestion immobilière
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePropertyBodyType } from "./createPropertyBodyType";

export interface CreatePropertyBody {
  title: string;
  description: string;
  type: CreatePropertyBodyType;
  address: string;
  city: string;
  country: string;
  price: number;
  /** @nullable */
  bedrooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  /** @nullable */
  area?: number | null;
  images: string[];
  /** @nullable */
  virtualTourUrl?: string | null;
  amenities?: string[];
  availableDates?: string[];
  isFeatured?: boolean;
}