# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * bloq5 — Plateforme de gestion immobilière
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyStatus } from "./propertyStatus";
import type { PropertyType } from "./propertyType";

export interface Property {
  id: number;
  title: string;
  description: string;
  type: PropertyType;
  address: string;
  city: string;
  country: string;
  price: number;
  /** @nullable */
  bedrooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  /** @nullable */
  area?: number | null;
  images: string[];
  /** @nullable */
  virtualTourUrl?: string | null;
  status: PropertyStatus;
  isFeatured: boolean;
  views: number;
  requestsCount: number;
  ownerId: string;
  createdAt: Date;
}